Add exclusion activity, minor export fix
To keep track of contacts from exclusion tests, this introduces
a new activity type "Exclusion Record". When creating activities
for a campaign, an activity of this type is created and assigned
to contacts marked for exclusion.

Additionally, deleting segments now causes the contacts to be
removed from the exclusion table.

This also includes a minor change to the segmentation export runner,
preventing empty (header-only) files from being generated.