Fix segmentation contact tab not rendering labels
This fixes an issue that's causing the segmentation tab to not render
labels for campaigns/segments/memberships. This is due to a change in
core where data tables are loaded via AJAX. The JS that should render
the labels executes before the table is actually rendered, so it's not
doing anything.